shop_product/serializers.py
from rest_framework import serializers
from .models import *


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        exclude = ('visit_count', 'favourite', 'sell')