requirements.txt
asgiref==3.3.1
boto3==1.17.16
botocore==1.20.16
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.6.0
Django==3.1.7
django-allauth==0.44.0
django-crispy-forms==1.11.1
django-filter==2.4.0
django-jalali==4.1.2
django-jazzmin==2.4.4
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
django-storages==1.11.1
djangorestframework==3.12.2
idna==2.10
jdatetime==3.6.2
jmespath==0.10.0
oauthlib==3.1.0
Pillow==8.1.0
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.8.0
pygount==1.2.4
PyJWT==2.0.1
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
suds-py3==1.4.4.1
urllib3==1.26.3
